+++ spec/factories/parent_technology_type.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :parent_technology_type, class: Patch::ParentTechnologyType do
+    slug { "forestry" }
+    name { "Forestry" }
+  end
+end