+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Adds mechanism, state, latitude, longitude, and technology_type to project responses
+- Adds mechanism, tagline, state, latitude, longitude, and technology_type to project responses
 
 ## [1.13.0] - 2021-09-10